+++ Editor/AvatarOptimizerSettings.cs
@@ -12,6 +12,12 @@ public class AvatarOptimizerSettings : EditorWindow
     private static d4rkAvatarOptimizer.Settings defaultSettings = new d4rkAvatarOptimizer.Settings();
     
     private static readonly string PrefsPrefix = "d4rkpl4y3r_AvatarOptimizer_";
+
+    public static bool DoOptimizeWithDefaultSettingsWhenNoComponent
+    {
+        get => EditorPrefs.GetBool(PrefsPrefix + "DoOptimizeWithDefaultSettingsWhenNoComponent", false);
+        private set => EditorPrefs.SetBool(PrefsPrefix + "DoOptimizeWithDefaultSettingsWhenNoComponent", value);
+    }
     
     [MenuItem("Tools/d4rkpl4y3r/Avatar Optimizer Settings")]
     static void Init()
@@ -24,9 +30,16 @@ public class AvatarOptimizerSettings : EditorWindow
     public void OnGUI()
     {
         scrollPos = EditorGUILayout.BeginScrollView(scrollPos);
+        BeginSection("Global Settings");
+        DoOptimizeWithDefaultSettingsWhenNoComponent = EditorGUILayout.ToggleLeft(
+            new GUIContent("Always Optimize on Upload", "If an Avatar does not have a d4rkAvatarOptimizer component attached, it will be optimized with the default settings below when uploading it to VRChat."),
+            DoOptimizeWithDefaultSettingsWhenNoComponent);
+        EndSection();
+        EditorGUILayout.Space();
+        BeginSection("Default Settings");
         EditorGUILayout.HelpBox("These settings are the default values when adding the Avatar Optimizer component to a model."
             +"You can change them here to suit your needs."
-            +"\nA * indicates settings that are different from their default value.", MessageType.Info);
+            +"\nA * indicates settings that are different from their default value.", MessageType.None);
         var fields = typeof(d4rkAvatarOptimizer.Settings).GetFields(System.Reflection.BindingFlags.Instance | System.Reflection.BindingFlags.Public)
             .Where(f => f.FieldType == typeof(bool) || f.FieldType == typeof(int)).ToArray();
         if (GUILayout.Button("Reset to Default"))
@@ -38,26 +51,46 @@ public class AvatarOptimizerSettings : EditorWindow
         }
         foreach (var field in fields)
         {
-            EditorGUILayout.BeginHorizontal();
-            bool isDefault = true;
             if (field.FieldType == typeof(bool))
             {
+                EditorGUILayout.BeginHorizontal();
                 GUILayoutUtility.GetRect(35, 15, GUILayout.Width(35));
-                var value = EditorGUILayout.Toggle(0 != GetValue(field.Name), GUILayout.Width(15)) ? 1 : 0;
+                var value = EditorGUILayout.ToggleLeft(d4rkAvatarOptimizer.GetDisplayName(field.Name) + (field.GetValue(defaultSettings).Equals(0 != GetValue(field.Name)) ? "" : " *"), 0 != GetValue(field.Name)) ? 1 : 0;
                 SetValue(field.Name, value);
-                isDefault = field.GetValue(defaultSettings).Equals(value != 0);
+                EditorGUILayout.EndHorizontal();
             }
             else
             {
+                EditorGUILayout.BeginHorizontal();
                 var value = EditorGUILayout.Popup(GetValue(field.Name), new string[] { "Off", "On", "Auto" }, GUILayout.Width(50));
                 SetValue(field.Name, value);
-                isDefault = field.GetValue(defaultSettings).Equals(value);
+                var content = new GUIContent(d4rkAvatarOptimizer.GetDisplayName(field.Name) + (field.GetValue(defaultSettings).Equals(value) ? "" : " *"));
+                var rect = GUILayoutUtility.GetRect(content, EditorStyles.label);
+                rect.xMin -= 2;
+                GUI.Label(rect, content);
+                EditorGUILayout.Space();
+                EditorGUILayout.EndHorizontal();
             }
-            EditorGUILayout.LabelField(d4rkAvatarOptimizer.GetDisplayName(field.Name) + (isDefault ? "" : " *"));
-            EditorGUILayout.EndHorizontal();
         }
-        EditorGUILayout.Space();
+        EndSection();
         EditorGUILayout.EndScrollView();
+    }
+
+    private void BeginSection(string title)
+    {
+        EditorGUILayout.BeginVertical(GUI.skin.box);
+        EditorGUILayout.LabelField(title, EditorStyles.boldLabel);
+        EditorGUILayout.BeginHorizontal();
+        GUILayoutUtility.GetRect(15, 15, GUILayout.Width(15));
+        EditorGUILayout.BeginVertical();
+    }
+
+    private void EndSection()
+    {
+        EditorGUILayout.EndVertical();
+        EditorGUILayout.EndHorizontal();
+        EditorGUILayout.Space();
+        EditorGUILayout.EndVertical();
     }
 
     public static int GetValue(string key)
@@ -67,11 +100,7 @@ public class AvatarOptimizerSettings : EditorWindow
         {
             throw new System.ArgumentException("Field " + key + " does not exist in d4rkAvatarOptimizer.Settings");
         }
-        if (!EditorPrefs.HasKey(PrefsPrefix + key))
-        {
-            EditorPrefs.SetInt(PrefsPrefix + key, field.FieldType == typeof(bool) ? (bool)field.GetValue(defaultSettings) ? 1 : 0 : (int)field.GetValue(defaultSettings));
-        }
-        return EditorPrefs.GetInt(PrefsPrefix + key);
+        return EditorPrefs.GetInt(PrefsPrefix + key, field.FieldType == typeof(bool) ? (bool)field.GetValue(defaultSettings) ? 1 : 0 : (int)field.GetValue(defaultSettings));
     }
 
     public static void SetValue(string key, int value)